Reimplement metadata queries with odbc metadata functions for pyodbc to avoid deadlocks (#94)
The metadata queries now use ODBC Metadata functions with pyodbc where possible to fetch the metadata or resolve names, because they are safe to use with not validated data. The ODBC metadata functions use snapshot metadata which avoid deadlocks on the metadata. As this it fixes issue #89  for pyodbc.
The previous implementation based on SQL Queries was renamed and kept as a fallback when the ODBC functions can't be used. For example, TurbODBC doesn't seem to provide them.
We added also tests for the metadata functions to ensure both implementations show the same behavior and the new implementation doesn't cause deadlocks anymore.