Add DBAPI2 compliant web-socket driver
In order to implement a web-socket based SQLA dialect for the Exasol-DB,
a DBAPI2 compliant web-socket driver is added. 

The current implementation of the driver is based on the `pyexasol` project
(If needed this can be changed in the future).

Additionally the unit and integration tests have been restructured
to simplify the test setup and execution. The restructuring is not completely done yet,
it rather should be seen as step towards the future layout/structure.